Kids Calculator Android Application
This is an Android Application that was created to be able to teach kids
about addition and subtraction and carry forward through actual
examples. The main feature of the app is its ability to explain every
problem in words, while using the specific numbers in the answer. The
app has a practice and a learn mode that utilizes the gyroscope to keep
kids engaged. This was an app that was written when I was teaching
myself Android Development.